% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SKMSD.R
\name{result_cv}
\alias{result_cv}
\title{Title}
\usage{
result_cv(r_SKMSD_cv, B = 1, filter_statistics = NULL, I = F)
}
\arguments{
\item{I}{}
}
\value{
SS1: Results based on the statistic of fliter in the method in the CV case
FP: Same as FDRPower;
SS_f1: results based on other fliter's statistics in the CV case.
}
\description{
Title
}
