% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SKMSD.R
\name{SKMSD}
\alias{SKMSD}
\title{Simultaneous knockoff for multi-source count data}
\usage{
SKMSD(
  W = W,
  class_K = class_K,
  data_x = NULL,
  M = NULL,
  y = y,
  T_var = 1,
  fdr = 0.2,
  offset = 1,
  test_statistic = "DE",
  filter_statistics = 3,
  test1 = "wilcox.test"
)
}
\arguments{
\item{W}{Count data}

\item{class_K}{Matrix data, representing the input of count data;}

\item{data_x}{Vector data, representing different sources of multi-source data; default is NULL, representing one source or pooling method;}

\item{M}{The sequencing depth of count data, which was the row sum of matrix data;}

\item{y}{Response variable, representing binary, continuous or mixed variables;}

\item{T_var}{True correlation variable, which represents the set of variables that are truly correlated with the response variable, defaults to NULL;}

\item{fdr}{the target FDR level, default 0.2;}

\item{offset}{value between 0 and 1;}

\item{test_statistic}{Statistics for a single source dataset, c('DE','GLM','RF')}

\item{filter_statistics}{The statistics of the dataset were calculated from multiple sources. Denote c(cumprod,max, sum) by 1,2,3;}

\item{test1}{Test method, select when selecting 'DE', default is "wilcox.test", c("wilcox.test",'ks.test','mmd','distance_JS').}
}
\value{
c_w: Results of test statistics for multiple single-source datasets;
c_w_b: The filter statisticss of the dataset were calculated from multiple sources;
S: The variable selection set;
FDRPower: The resulting FDR and Power are calculated after given T_var.
}
\description{
Simultaneous knockoff for multi-source count data
}
