% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SKMSD.R
\name{SKMSD_B}
\alias{SKMSD_B}
\title{An Aggregating version of SKMSD}
\usage{
SKMSD_B(
  W = W,
  class_K = NULL,
  data_x = NULL,
  M = NULL,
  y = y,
  T_var = NULL,
  fdr = 0.2,
  offset = 1,
  B = 1,
  Bstat = 2,
  test_statistic = "DE",
  filter_statistics = 3,
  test1 = "wilcox.test",
  combine_1 = "simul"
)
}
\arguments{
\item{W}{Same as SKMSD;}

\item{B}{Number of aggregations;}

\item{Bstat}{For aggregation, 1 represents the aggregated test statistic and 2 represents the aggregated fliter statistic;}

\item{combine_1}{When combine_1 aggregation mode is 1, 'simul' means that the statistic is calculated based on the proposed method, and 'fisher' means that the statistic is calculated based on the combination of p values.}
}
\value{
S: The variable selection set;
c_w_pis: Same as c_w;
contrast_score: fliter statistics for the non-aggregated case (B=1);
FDRPower: The resulting FDR and Power are calculated after given T_var.
chis: The threshold of the chi-square distribution using fisher's method in the first aggregation case (Bstat = 1, combine_1 == 'fisher');
p_comb:
e_value: Combined E-values for multi-source datasets;
e_w_B: The e value of the multi-source dataset in the aggregated case, B times;
e_value: E-values for single-source datasets.
}
\description{
An Aggregating version of SKMSD
}
