% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SKMSD.R
\name{SKMSD_B_other}
\alias{SKMSD_B_other}
\title{Extension of SKMSD_B to consider more synthetic data methods.}
\usage{
SKMSD_B_other(
  W = W,
  class_K = NULL,
  data_x = NULL,
  M = NULL,
  y = y,
  T_var = NULL,
  fdr = 0.2,
  offset = 1,
  method = "ZIPG",
  B = 1,
  Bstat = 2,
  test_statistic = "DE",
  filter_statistics = 3,
  test1 = "wilcox.test",
  combine_1 = "simul"
)
}
\arguments{
\item{W}{Same as SKMSD_B;}

\item{method}{Synthetic data generation methods, c('ZIPG','ZINB','knockoff');}

\item{combine_1}{}
}
\value{
Same as SKMSD_B;
}
\description{
Extension of SKMSD_B to consider more synthetic data methods.
}
