% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SKMSD.R
\name{SKMSD_cv}
\alias{SKMSD_cv}
\title{SKMSD method based on the cross-validation case}
\usage{
SKMSD_cv(
  W = W,
  class_K = NULL,
  data_x = NULL,
  M = NULL,
  y = y,
  T_var = NULL,
  fdr = 0.2,
  offset = 1,
  cv = 1,
  method = "ZIPG",
  B = 1,
  Bstat = 2,
  test_statistic = "DE",
  filter_statistics = 3,
  test1 = "wilcox.test",
  combine_1 = "simul",
  I = F
)
}
\arguments{
\item{W}{Same as SKMSD;}

\item{cv}{Fold number of cross validation;}

\item{method}{Synthetic data generation methods, c('ZIPG','ZINB','knockoff');}

\item{I}{Boolean value that represents the result of the computation based on the Intersection.}
}
\value{
Same as SKMSD and SKMSD_B;
}
\description{
SKMSD method based on the cross-validation case
}
