% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SKMSD.R
\name{SKMSD_other}
\alias{SKMSD_other}
\title{Extension of SKMSD to consider more synthetic data methods.}
\usage{
SKMSD_other(
  W = W,
  class_K = NULL,
  data_x = NULL,
  M = NULL,
  y = y,
  T_var = NULL,
  fdr = 0.2,
  offset = 1,
  method = "ZIPG",
  test_statistic = "DE",
  filter_statistics = 3,
  test1 = "wilcox.test"
)
}
\arguments{
\item{W}{Same as SKMSD;}

\item{method}{Synthetic data generation methods, c('ZIPG','ZINB','knockoff');}

\item{test1}{}
}
\value{
Same as SKMSD;
}
\description{
Extension of SKMSD to consider more synthetic data methods.
}
